requirements.txt
requests>=2.31.0
openai>=1.0.0
plotly>=5.0.0
geopy>=2.4.0
Pillow>=10.0.0